+++ mzLib/Test/Proteomics/ProteolyticDigestion/PeptideEqualityAndCloneTests.cs
@@ -1,0 +1,197 @@
+// Copyright 2012, 2013, 2014 Derek J. Bailey
+// Modified work copyright 2016 Stefan Solntsev
+//
+// This file (PeptideEqualityAndCloneTests.cs) is part of Proteomics.
+//
+// Proteomics is free software: you can redistribute it and/or modify it
+// under the terms of the GNU Lesser General Public License as published
+// by the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Proteomics is distributed in the hope that it will be useful, but WITHOUT
+// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+// FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public
+// License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public
+// License along with Proteomics. If not, see <http://www.gnu.org/licenses/>.
+
+using Chemistry;
+using NUnit.Framework;
+using Assert = NUnit.Framework.Legacy.ClassicAssert;
+using Proteomics.AminoAcidPolymer;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using Stopwatch = System.Diagnostics.Stopwatch;
+
+namespace Test
+{
+    [TestFixture]
+    [System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
+    public sealed class PeptideEqualityAndCloneTests
+    {
+        private Peptide _mockPeptideEveryAminoAcid;
+        private Peptide _mockTrypticPeptide;
+        private static Stopwatch Stopwatch { get; set; }
+
+        [SetUp]
+        public static void Setuppp()
+        {
+            Stopwatch = new Stopwatch();
+            Stopwatch.Start();
+        }
+
+        [TearDown]
+        public static void TearDown()
+        {
+            Console.WriteLine($"Analysis time: {Stopwatch.Elapsed.Hours}h {Stopwatch.Elapsed.Minutes}m {Stopwatch.Elapsed.Seconds}s");
+        }
+
+        [SetUp]
+        public void SetUp()
+        {
+            _mockPeptideEveryAminoAcid = new Peptide("ACDEFGHIKLMNPQRSTVWY");
+            _mockTrypticPeptide = new Peptide("TTGSSSSSSSK");
+        }
+
+        [Test]
+        public void AApolymerNullEquals()
+        {
+            Peptide pep = new Peptide("G");
+            Assert.IsFalse(pep.Equals(null));
+        }
+
+        [Test]
+        public void PeptideEquality()
+        {
+            Peptide pepA = new Peptide("DEREK");
+            Peptide pepB = new Peptide("DEREK");
+            Assert.AreEqual(pepA, pepB);
+
+            Peptide pepC = new Peptide("DEREKK");
+            Assert.AreNotEqual(pepA, pepC);
+        }
+
+        [Test]
+        public void PeptideInEqualityAminoAcidSwitch()
+        {
+            Peptide pepA = new Peptide("DEREK");
+            Peptide pepB = new Peptide("DEERK");
+            Assert.AreNotEqual(pepA, pepB);
+        }
+
+        [Test]
+        public void PeptideInEqualityAminoAcidModification()
+        {
+            Peptide pepA = new Peptide("DEREK");
+            Peptide pepB = new Peptide("DEREK");
+            pepB.SetModification(new OldSchoolChemicalFormulaModification(ChemicalFormula.ParseFormula("H2O")), 'R');
+            Assert.AreNotEqual(pepA, pepB);
+            pepA.SetModification(new OldSchoolChemicalFormulaModification(ChemicalFormula.ParseFormula("H2O2")), 'R');
+            Assert.AreNotEqual(pepA, pepB);
+        }
+
+        [Test]
+        public void PeptideCloneEquality()
+        {
+            Peptide pepA = new Peptide("DEREK");
+            Peptide pepB = new Peptide(pepA);
+            Assert.AreEqual(pepA, pepB);
+        }
+
+        [Test]
+        public void PeptideCloneNotSameReference()
+        {
+            Peptide pepA = new Peptide("DEREK");
+            Peptide pepB = new Peptide(pepA);
+
+            Assert.AreNotSame(pepA, pepB);
+        }
+
+        [Test]
+        public void PeptideCloneWithModifications()
+        {
+            Peptide pepA = new Peptide("DER[Fe]EK");
+            Peptide pepB = new Peptide(pepA);
+            Assert.AreEqual("DER[Fe]EK", pepB.ToString());
+        }
+
+        [Test]
+        public void PeptideCloneWithoutModifications()
+        {
+            Peptide pepA = new Peptide("DER[Fe]EK");
+            Peptide pepB = new Peptide(pepA, false);
+            Assert.AreEqual("DEREK", pepB.ToString());
+        }
+
+        [Test]
+        public void PeptideCloneWithModification()
+        {
+            Peptide pepA = new Peptide("DEREK");
+            pepA.SetModification(new OldSchoolChemicalFormulaModification(ChemicalFormula.ParseFormula("H2O")), 'R');
+            Peptide pepB = new Peptide(pepA);
+
+            Assert.AreEqual(pepB, pepA);
+        }
+
+        [Test]
+        public void PeptideParitalCloneInternal()
+        {
+            Peptide pepA = new Peptide("DEREK");
+            Peptide pepB = new Peptide(pepA, 1, 3);
+            Peptide pepC = new Peptide("ERE");
+            Assert.AreEqual(pepB, pepC);
+        }
+
+        [Test]
+        public void PeptideParitalClonelWithInternalModification()
+        {
+            Peptide pepA = new Peptide("DER[Fe]EK");
+            Peptide pepB = new Peptide(pepA, 2, 3);
+            Peptide pepC = new Peptide("R[Fe]EK");
+            Assert.AreEqual(pepB, pepC);
+        }
+
+        [Test]
+        public void PeptideParitalClonelWithInternalModificationTwoMods()
+        {
+            Peptide pepA = new Peptide("DE[Al]R[Fe]EK");
+            Peptide pepB = new Peptide(pepA, 2, 3);
+            Peptide pepC = new Peptide("R[Fe]EK");
+            Assert.AreEqual(pepB, pepC);
+        }
+
+        [Test]
+        public void PeptideParitalCloneInternalWithCTerminusModification()
+        {
+            Peptide pepA = new Peptide("DEREK");
+            pepA.SetModification(new OldSchoolChemicalFormulaModification(ChemicalFormula.ParseFormula("H2O")), Terminus.C);
+            Peptide pepB = new Peptide(pepA, 2, 3);
+
+            Peptide pepC = new Peptide("REK");
+            pepC.SetModification(new OldSchoolChemicalFormulaModification(ChemicalFormula.ParseFormula("H2O")), Terminus.C);
+
+            Assert.AreEqual(pepC, pepB);
+        }
+
+        [Test]
+        public void PeptideHashing()
+        {
+            Peptide pep1 = new Peptide("DEREK");
+            Peptide pep2 = new Peptide("DEREKN");
+            Peptide pep3 = new Peptide("DEREKM");
+            Peptide pep4 = new Peptide("DEREKM");
+            HashSet<Peptide> uu = new HashSet<Peptide> { pep1, pep2, pep3, pep4 };
+            uu.Add(new Peptide("DEREKN"));
+            Assert.AreEqual(3, uu.Count);
+        }
+
+        [Test]
+        public void TestGetSubPeptide()
+        {
+            Peptide pep = new Peptide("DERLEK");
+            Assert.AreEqual(new Peptide("LE"), pep.GetSubPeptide(3, 2));
+        }
+    }
+}
